\name{ARFIMA-class}
\Rdversion{1.1}
\docType{class}
\alias{ARFIMA-class}
\title{class: High Level ARFIMA class}
\description{
The virtual parent class of the ARFIMA subset.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Extends}{
Class \code{"\linkS4class{rGARCH}"}, directly.
}
\section{Methods}{
No methods defined with class "ARFIMA" in the signature.
}
\author{ Alexios Ghalanos }
\keyword{classes}
