\name{ARFIMAmultispec-class}
\Rdversion{1.1}
\docType{class}
\alias{ARFIMAmultispec-class}
\alias{show,ARFIMAmultispec-method}
\title{class: ARFIMA Multiple Specification Class}
\description{
Class for the ARFIMA Multiple specification.}
\section{Slots}{
  \describe{
    \item{\code{spec}:}{Object of class \code{"vector"} }
    \item{\code{type}:}{Object of class \code{"character"} }
  }
}
\section{Extends}{
Class \code{"\linkS4class{ARFIMA}"}, directly.
Class \code{"\linkS4class{rGARCH}"}, by class "ARFIMA", distance 2.
}
\section{Methods}{
  \describe{
    \item{show}{\code{signature(object = "ARFIMAmultispec")}: specification summary.}
	 }
}
\author{Alexios Ghalanos}
\keyword{classes}
