\name{GARCHboot-class}
\docType{class}
\alias{GARCHboot-class}
\title{class: GARCH Bootstrap Class}
\description{High Level GARCH bootstrap class to hold the univariate and multivariate boot
objects.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Extends}{
Class \code{"\linkS4class{rGARCH}"}, directly.
}
\section{Methods}{
No methods defined with class "GARCHboot" in the signature.
}
\author{Alexios Ghalanos}
\examples{
showClass("GARCHboot")
}
\keyword{classes}