\name{GARCHdistribution-class}
\docType{class}
\alias{GARCHdistribution-class}
\title{class: GARCH Parameter Distribution Class}
\description{High Level GARCH parameter distribution class to hold the univariate 
and multivariate boot objects.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Extends}{
Class \code{"\linkS4class{rGARCH}"}, directly.
}
\section{Methods}{
No methods defined with class "GARCHdistribution" in the signature.
}
\author{Alexios Ghalanos}
\examples{
showClass("GARCHdistribution")
}
\keyword{classes}