\name{GARCHfilter-class}
\docType{class}
\alias{GARCHfilter-class}
\title{class: GARCH Filter Class}
\description{High Level GARCH filter class to hold the univariate and multivariate filter
objects.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Extends}{
Class \code{"\linkS4class{rGARCH}"}, directly.
}
\section{Methods}{
No methods defined with class "GARCHfilter" in the signature.
}
\author{Alexios Ghalanos}
\examples{
showClass("GARCHfilter")
}
\keyword{classes}