\name{GARCHfit-class}
\docType{class}
\alias{GARCHfit-class}
\title{class: GARCH Fit Class}
\description{High Level GARCH fit class to hold the univariate and multivariate fits
objects.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Extends}{
Class \code{"\linkS4class{rGARCH}"}, directly.
}
\section{Methods}{
No methods defined with class "GARCHfit" in the signature.
}
\author{Alexios Ghalanos}
\examples{
showClass("GARCHfit")
}
\keyword{classes}