\name{GARCHforecast-class}
\docType{class}
\alias{GARCHforecast-class}
\title{class: GARCH Forecast Class}
\description{High Level GARCH forecast class to hold the univariate and multivariate 
forecast objects.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Extends}{
Class \code{"\linkS4class{rGARCH}"}, directly.
}
\section{Methods}{
No methods defined with class "GARCHforecast" in the signature.
}
\author{Alexios Ghalanos}
\examples{
showClass("GARCHforecast")
}
\keyword{classes}
