\name{GARCHpath-class}
\docType{class}
\alias{GARCHpath-class}
\title{class: GARCH Path Simulation Class}
\description{High Level GARCH Path simulation class to hold the univariate and multivariate 
path simulation objects.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Extends}{
Class \code{"\linkS4class{rGARCH}"}, directly.
}
\section{Methods}{
No methods defined with class "GARCHpath" in the signature.
}
\author{Alexios Ghalanos}
\examples{
showClass("GARCHpath")
}
\keyword{classes}
