\name{GARCHroll-class}
\docType{class}
\alias{GARCHroll-class}
\title{class: GARCH Roll Class}
\description{High Level GARCH roll class to hold the univariate and multivariate roll
objects.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Extends}{
Class \code{"\linkS4class{rGARCH}"}, directly.
}
\section{Methods}{
No methods defined with class "GARCHroll" in the signature.
}
\author{Alexios Ghalanos}
\examples{
showClass("GARCHroll")
}
\keyword{classes}