\name{GARCHsim-class}
\docType{class}
\alias{GARCHsim-class}
\title{class: GARCH Simulation Class}
\description{High Level GARCH simulation class to hold the univariate and multivariate 
simulation objects.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Extends}{
Class \code{"\linkS4class{rGARCH}"}, directly.
}
\section{Methods}{
No methods defined with class "GARCHsim" in the signature.
}
\author{Alexios Ghalanos}
\examples{
showClass("GARCHsim")
}
\keyword{classes}
