\name{GARCHspec-class}
\docType{class}
\alias{GARCHspec-class}
\title{class: GARCH Spec Class}
\description{High Level GARCH spec class to hold the univariate and multivariate 
spec objects.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Extends}{
Class \code{"\linkS4class{rGARCH}"}, directly.
}
\section{Methods}{
No methods defined with class "GARCHspec" in the signature.
}
\author{Alexios Ghalanos}
\examples{
showClass("GARCHspec")
}
\keyword{classes}
