\name{GARCHtests-class}
\Rdversion{1.1}
\docType{class}
\alias{GARCHtests-class}
\title{class: GARCH Tests Class}
\description{GARCH High level inference and other tests class.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Extends}{
Class \code{"\linkS4class{rGARCH}"}, directly.
}
\section{Methods}{
No methods defined with class "GARCHtests" in the signature.
}
\author{Alexios Ghalanos}
\examples{
showClass("GARCHtests")
}
\keyword{classes}