\name{rGARCH-class}
\docType{class}
\alias{rGARCH-class}
\title{class: rGARCH Class}
\description{
Highest Level Virtual Package Class to which all other classes belong.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "rGARCH" in the signature.
}
\examples{
showClass("rGARCH")
}
\author{Alexios Ghalanos}
\keyword{classes}
