\name{sp500ret}
\docType{data}
\alias{sp500ret}
\title{data: Standard and Poors 500 Closing Value Log Return}
\description{
The S\&P500 index closing value log return from 1987-03-10 to 2009-01-30 from
yahoo finance.
}
\usage{data(sp500ret)}
\format{A data.frame containing 1x5523 observations.}
\source{Yahoo Finance}
\keyword{datasets}
