\name{spyreal}
\docType{data}
\alias{spyreal}
\title{data: SPDR Standard and Poors 500 Open-Close Daily Return and Realized Kernel Volatility}
\description{
The SPDR SP500 index open-close return and the realized kernel volatility for the period 
2002-01-02 to 2008-08-29 from the paper of Hansen, Huang and Shek (2011). Used for illustrating
the implementation of the Realized GARCH model in rugarch.
}
\usage{data(spyreal)}
\format{An xts object.}
\source{Journal of Applied Econometrics Data Archive}
\references{
Hansen, P. R., Huang, Z., and Shek, H. H. (2012). Realized GARCH: a joint model for returns 
and realized measures of volatility. \emph{Journal of Applied Econometrics}, \bold{27(6)}, 
877--906.
}
\keyword{datasets}
