\name{uGARCHmultispec-class}
\docType{class}
\alias{show,uGARCHmultispec-method}
\alias{uGARCHmultispec-class}
\title{class: Univariate GARCH Multiple Specification Class}
\description{Class for the univariate GARCH Multiple specification.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Extends}{
Class \code{"\linkS4class{GARCHspec}"}, directly.
Class \code{"\linkS4class{rGARCH}"}, by class "GARCHspec", distance 3.
}
\section{Methods}{
  \describe{
  \item{show}{\code{signature(object = "uGARCHmultispec")}: specification summary. }
  }
}
\author{Alexios Ghalanos}
\seealso{
Classes \code{\linkS4class{uGARCHmultiforecast}}, \code{\linkS4class{uGARCHmultifit}} and \code{\linkS4class{uGARCHmultifilter}}.
}
\keyword{classes}